% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifylinear.R
\name{classify.linear}
\alias{classify.linear}
\title{Classify Linear}
\usage{
classify.linear(x, w, b)
}
\arguments{
\item{x}{is vector of inputs}

\item{w}{is a unit vector}

\item{b}{is the directed distance b}
}
\description{
Classify Linear
}
