% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancefromplan.R
\name{distance.from.plan}
\alias{distance.from.plan}
\title{Distance from plan}
\usage{
distance.from.plan(z, w, b)
}
\arguments{
\item{z}{is a vector of x}

\item{w}{is a unit vector}

\item{b}{is the directed distance b}
}
\description{
Given any seperator, the idea that one side of the seperator
are predicte belong to other class, while other side are predicted
to belong to a different class
}
