% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{perceptron}
\alias{perceptron}
\title{Simple Perceptron Algorithm}
\usage{
perceptron(x, y, learning.rate = 1)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{learning.rate}{intialialize learning rate to 1}
}
\description{
Simple Perceptron Algorithm
}
