% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{simplePerceptron}
\alias{simplePerceptron}
\title{Plot the data subset graph and error graph of sample dataset}
\usage{
simplePerceptron(datasub, xtrain, ytarget, weight, learningRate)
}
\arguments{
\item{datasub}{random sample from training set}

\item{xtrain}{Feature vector}

\item{ytarget}{class label of sample}

\item{weight}{Initial weight vector}

\item{learningRate}{learning Rate}
}
\value{
weight outputs weight vector and plots classifier line
}
\description{
Takes in any subset
}
\author{
Amitsingh Pardeshi
# @param func Pass in the function type (step, sigmoid, relu)
}
